requirements.txt
flask
pyMuPDF
python-docx
werkzeug